rename use-new-note
